This patch requires a cache_node table (and is currently dependent on memcache, though there's no reason it couldn't also work with MySQL or another cache).
This patch only caches the latest revision of each node.  It would be
possible to support older revisions too if used heavily, but as that seems
unlikely this patch currently avoids the complexity.

The patch only caches polls for anonymous users.

The cache_node table never gets cleared with a wildcard, so it's only
safe to combine it with another bin that also doesn't get flushed.